worker/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
aio-pika
pytest
httpx
pytest-asyncio